refactor(settings_ui): extract file persistence to _persistence.py
Config/metadata-cache/backup-override file I/O + the atomic writer and
override-file lock. Decomposes load_tool_config (C901 14 -> under
threshold) via _seed_tool_config_from_env. Handlers reach the helpers via
_persistence.<name>; test patch targets and tools_dev imports repointed to
the canonical ha_mcp.settings_ui._persistence.<name> site.
